企业微信部门创建接口create返回值改为long类型

+++ weixin-java-cp/src/main/java/me/chanjar/weixin/cp/api/WxCpDepartmentService.java
@@ -26,7 +26,7 @@ public interface WxCpDepartmentService {
    * @return 部门id
    * @throws WxErrorException 异常
    */
-  Integer create(WxCpDepart depart) throws WxErrorException;
+  Long create(WxCpDepart depart) throws WxErrorException;
 
   /**
    * <pre>
